refactor(prisma): re-export prof_role from db singleton
